--- a/src/main/java/io/vertx/core/http/impl/Http1xClientConnection.java
+++ b/src/main/java/io/vertx/core/http/impl/Http1xClientConnection.java
@@ -414,7 +414,7 @@ public class Http1xClientConnection extends Http1xConnection implements HttpClie
       super(context, promise, id);
 
       this.conn = conn;
-      this.queue = new InboundMessageQueue<>(context) {
+      this.queue = new InboundMessageQueue<>(conn.context.nettyEventLoop(), context) {
         @Override
         protected void handleResume() {
           conn.doResume();