--- a/src/main/java/io/vertx/core/http/impl/Http1xClientConnection.java
+++ b/src/main/java/io/vertx/core/http/impl/Http1xClientConnection.java
@@ -424,7 +424,7 @@ public class Http1xClientConnection extends Http1xConnection implements HttpClie
           conn.doPause();
         }
         @Override
-        protected void handle(Object item) {
+        protected void handleMessage(Object item) {
           if (!reset) {
             if (item instanceof MultiMap) {
               Handler<MultiMap> handler = endHandler;