--- a/src/main/java/io/vertx/core/http/impl/Http1xServerConnection.java
+++ b/src/main/java/io/vertx/core/http/impl/Http1xServerConnection.java
@@ -80,6 +80,7 @@ public class Http1xServerConnection extends Http1xConnection implements HttpServ
 
   private Http1xServerRequest requestInProgress;
   private Http1xServerRequest responseInProgress;
+  private Http1xServerRequest pipelinedRequest;
   private boolean wantClose;
   private Handler<HttpServerRequest> requestHandler;
   private Handler<HttpServerRequest> invalidRequestHandler;
@@ -154,7 +155,8 @@ public class Http1xServerConnection extends Http1xConnection implements HttpServ
       Http1xServerRequest req = new Http1xServerRequest(this, request, requestCtx, parked);
       requestInProgress = req;
       if (parked) {
-        enqueueRequest(req);
+        pipelinedRequest = req;
+        doPause();
         return;
       }
       boolean keepAlive = HttpUtils.isKeepAlive(request);
@@ -166,11 +168,6 @@ public class Http1xServerConnection extends Http1xConnection implements HttpServ
     } else {
       handleOther(msg);
     }
-  }
-
-  private void enqueueRequest(Http1xServerRequest req) {
-    // Deferred until the current response completion
-    responseInProgress.enqueue(req);
   }
 
   private void handleOther(Object msg) {
@@ -235,8 +232,9 @@ public class Http1xServerConnection extends Http1xConnection implements HttpServ
         if (requestInProgress == request) {
           // Deferred
         } else {
-          Http1xServerRequest next = request.next();
+          Http1xServerRequest next = pipelinedRequest;
           if (next != null) {
+            pipelinedRequest = null;
             // Handle pipelined request
             handleNext(next);
           } else if (wantClose || shutdownInitiated) {
@@ -258,11 +256,12 @@ public class Http1xServerConnection extends Http1xConnection implements HttpServ
     responseInProgress = next;
     wantClose |= !keepAlive;
     next.handleBegin(keepAlive);
+    next.parked = false;
     next.context.emit(next, next_ -> {
       Handler<HttpServerRequest> handler = next_.nettyRequest().decoderResult().isSuccess() ? requestHandler : invalidRequestHandler;
       handler.handle(next_);
-      next_.unpark();
     });
+    doResume();
   }
 
   private void reportResponseComplete() {