--- a/src/main/java/io/vertx/core/http/impl/Http1xServerConnection.java
+++ b/src/main/java/io/vertx/core/http/impl/Http1xServerConnection.java
@@ -152,7 +152,7 @@ public class Http1xServerConnection extends Http1xConnection implements HttpServ
       DefaultHttpRequest request = (DefaultHttpRequest) msg;
       ContextInternal requestCtx = streamContextSupplier.get();
       boolean parked = responseInProgress != null;
-      Http1xServerRequest req = new Http1xServerRequest(this, request, requestCtx, parked);
+      Http1xServerRequest req = new Http1xServerRequest(this, request, requestCtx);
       requestInProgress = req;
       if (parked) {
         pipelinedRequest = req;
@@ -256,7 +256,6 @@ public class Http1xServerConnection extends Http1xConnection implements HttpServ
     responseInProgress = next;
     wantClose |= !keepAlive;
     next.handleBegin(keepAlive);
-    next.parked = false;
     next.context.emit(next, next_ -> {
       Handler<HttpServerRequest> handler = next_.nettyRequest().decoderResult().isSuccess() ? requestHandler : invalidRequestHandler;
       handler.handle(next_);