--- a/src/main/java/io/vertx/core/http/impl/Http1xServerConnection.java
+++ b/src/main/java/io/vertx/core/http/impl/Http1xServerConnection.java
@@ -151,10 +151,10 @@ public class Http1xServerConnection extends Http1xConnection implements HttpServ
       // fast path type check vs concrete class
       DefaultHttpRequest request = (DefaultHttpRequest) msg;
       ContextInternal requestCtx = streamContextSupplier.get();
-      boolean parked = responseInProgress != null;
       Http1xServerRequest req = new Http1xServerRequest(this, request, requestCtx);
       requestInProgress = req;
-      if (parked) {
+      if (responseInProgress != null) {
+        assert pipelinedRequest == null;
         pipelinedRequest = req;
         doPause();
         return;